More refactoring of tests into individual test files fr each math/trig function
