+++ tests/PhpSpreadsheetTests/Calculation/Functions/MathTrig/CombinTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\MathTrig;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\MathTrig;
+use PHPUnit\Framework\TestCase;
+
+class CombinTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerCOMBIN
+     *
+     * @param mixed $expectedResult
+     */
+    public function testCOMBIN($expectedResult, ...$args)
+    {
+        $result = MathTrig::COMBIN(...$args);
+        $this->assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerCOMBIN()
+    {
+        return require 'data/Calculation/MathTrig/COMBIN.php';
+    }
+}
