+++ tests/PhpSpreadsheetTests/Calculation/Functions/MathTrig/EvenTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\MathTrig;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\MathTrig;
+use PHPUnit\Framework\TestCase;
+
+class EvenTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerEVEN
+     *
+     * @param mixed $expectedResult
+     */
+    public function testEVEN($expectedResult, ...$args)
+    {
+        $result = MathTrig::EVEN(...$args);
+        $this->assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerEVEN()
+    {
+        return require 'data/Calculation/MathTrig/EVEN.php';
+    }
+}
