+++ tests/PhpSpreadsheetTests/Calculation/Functions/MathTrig/FactTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\MathTrig;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\MathTrig;
+use PHPUnit\Framework\TestCase;
+
+class FactTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerFACT
+     *
+     * @param mixed $expectedResult
+     */
+    public function testFACT($expectedResult, ...$args)
+    {
+        $result = MathTrig::FACT(...$args);
+        $this->assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerFACT()
+    {
+        return require 'data/Calculation/MathTrig/FACT.php';
+    }
+}
