+++ tests/PhpSpreadsheetTests/Calculation/Functions/MathTrig/FloorTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\MathTrig;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\MathTrig;
+use PHPUnit\Framework\TestCase;
+
+class FloorTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerFLOOR
+     *
+     * @param mixed $expectedResult
+     */
+    public function testFLOOR($expectedResult, ...$args)
+    {
+        $result = MathTrig::FLOOR(...$args);
+        $this->assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerFLOOR()
+    {
+        return require 'data/Calculation/MathTrig/FLOOR.php';
+    }
+}
