+++ tests/PhpSpreadsheetTests/Calculation/Functions/MathTrig/IntTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\MathTrig;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\MathTrig;
+use PHPUnit\Framework\TestCase;
+
+class IntTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerINT
+     *
+     * @param mixed $expectedResult
+     */
+    public function testINT($expectedResult, ...$args)
+    {
+        $result = MathTrig::INT(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerINT()
+    {
+        return require 'data/Calculation/MathTrig/INT.php';
+    }
+}
