+++ tests/PhpSpreadsheetTests/Calculation/Functions/MathTrig/SignTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\MathTrig;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\MathTrig;
+use PHPUnit\Framework\TestCase;
+
+class SignTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerSIGN
+     *
+     * @param mixed $expectedResult
+     */
+    public function testSIGN($expectedResult, ...$args)
+    {
+        $result = MathTrig::SIGN(...$args);
+        $this->assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerSIGN()
+    {
+        return require 'data/Calculation/MathTrig/SIGN.php';
+    }
+}
