+++ tests/PhpSpreadsheetTests/Calculation/MathTrigTest.php
@@ -18,246 +18,6 @@ class MathTrigTest extends TestCase
     }
 
     /**
-     * @dataProvider providerATAN2
-     *
-     * @param mixed $expectedResult
-     */
-    public function testATAN2($expectedResult, ...$args)
-    {
-        $result = MathTrig::ATAN2(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerATAN2()
-    {
-        return require 'data/Calculation/MathTrig/ATAN2.php';
-    }
-
-    /**
-     * @dataProvider providerCEILING
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCEILING($expectedResult, ...$args)
-    {
-        $result = MathTrig::CEILING(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerCEILING()
-    {
-        return require 'data/Calculation/MathTrig/CEILING.php';
-    }
-
-    /**
-     * @dataProvider providerCOMBIN
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOMBIN($expectedResult, ...$args)
-    {
-        $result = MathTrig::COMBIN(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerCOMBIN()
-    {
-        return require 'data/Calculation/MathTrig/COMBIN.php';
-    }
-
-    /**
-     * @dataProvider providerEVEN
-     *
-     * @param mixed $expectedResult
-     */
-    public function testEVEN($expectedResult, ...$args)
-    {
-        $result = MathTrig::EVEN(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerEVEN()
-    {
-        return require 'data/Calculation/MathTrig/EVEN.php';
-    }
-
-    /**
-     * @dataProvider providerODD
-     *
-     * @param mixed $expectedResult
-     */
-    public function testODD($expectedResult, ...$args)
-    {
-        $result = MathTrig::ODD(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerODD()
-    {
-        return require 'data/Calculation/MathTrig/ODD.php';
-    }
-
-    /**
-     * @dataProvider providerFACT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testFACT($expectedResult, ...$args)
-    {
-        $result = MathTrig::FACT(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerFACT()
-    {
-        return require 'data/Calculation/MathTrig/FACT.php';
-    }
-
-    /**
-     * @dataProvider providerFACTDOUBLE
-     *
-     * @param mixed $expectedResult
-     */
-    public function testFACTDOUBLE($expectedResult, ...$args)
-    {
-        $result = MathTrig::FACTDOUBLE(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerFACTDOUBLE()
-    {
-        return require 'data/Calculation/MathTrig/FACTDOUBLE.php';
-    }
-
-    /**
-     * @dataProvider providerFLOOR
-     *
-     * @param mixed $expectedResult
-     */
-    public function testFLOOR($expectedResult, ...$args)
-    {
-        $result = MathTrig::FLOOR(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerFLOOR()
-    {
-        return require 'data/Calculation/MathTrig/FLOOR.php';
-    }
-
-    /**
-     * @dataProvider providerGCD
-     *
-     * @param mixed $expectedResult
-     */
-    public function testGCD($expectedResult, ...$args)
-    {
-        $result = MathTrig::GCD(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerGCD()
-    {
-        return require 'data/Calculation/MathTrig/GCD.php';
-    }
-
-    /**
-     * @dataProvider providerLCM
-     *
-     * @param mixed $expectedResult
-     */
-    public function testLCM($expectedResult, ...$args)
-    {
-        $result = MathTrig::LCM(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerLCM()
-    {
-        return require 'data/Calculation/MathTrig/LCM.php';
-    }
-
-    /**
-     * @dataProvider providerINT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testINT($expectedResult, ...$args)
-    {
-        $result = MathTrig::INT(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerINT()
-    {
-        return require 'data/Calculation/MathTrig/INT.php';
-    }
-
-    /**
-     * @dataProvider providerSIGN
-     *
-     * @param mixed $expectedResult
-     */
-    public function testSIGN($expectedResult, ...$args)
-    {
-        $result = MathTrig::SIGN(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerSIGN()
-    {
-        return require 'data/Calculation/MathTrig/SIGN.php';
-    }
-
-    /**
-     * @dataProvider providerPOWER
-     *
-     * @param mixed $expectedResult
-     */
-    public function testPOWER($expectedResult, ...$args)
-    {
-        $result = MathTrig::POWER(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerPOWER()
-    {
-        return require 'data/Calculation/MathTrig/POWER.php';
-    }
-
-    /**
-     * @dataProvider providerLOG
-     *
-     * @param mixed $expectedResult
-     */
-    public function testLOG($expectedResult, ...$args)
-    {
-        $result = MathTrig::logBase(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerLOG()
-    {
-        return require 'data/Calculation/MathTrig/LOG.php';
-    }
-
-    /**
-     * @dataProvider providerMOD
-     *
-     * @param mixed $expectedResult
-     */
-    public function testMOD($expectedResult, ...$args)
-    {
-        $result = MathTrig::MOD(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerMOD()
-    {
-        return require 'data/Calculation/MathTrig/MOD.php';
-    }
-
-    /**
      * @dataProvider providerMDETERM
      *
      * @param mixed $expectedResult
@@ -303,22 +63,6 @@ class MathTrigTest extends TestCase
     public function providerMMULT()
     {
         return require 'data/Calculation/MathTrig/MMULT.php';
-    }
-
-    /**
-     * @dataProvider providerMULTINOMIAL
-     *
-     * @param mixed $expectedResult
-     */
-    public function testMULTINOMIAL($expectedResult, ...$args)
-    {
-        $result = MathTrig::MULTINOMIAL(...$args);
-        self::assertEquals($expectedResult, $result, null, 1E-12);
-    }
-
-    public function providerMULTINOMIAL()
-    {
-        return require 'data/Calculation/MathTrig/MULTINOMIAL.php';
     }
 
     /**
